ai_kasetFair/requirement.txt
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
cryptography==46.0.3
debugpy==1.8.17
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
executing==2.2.1
fastapi==0.115.5
fsspec==2025.12.0
gotrue==2.9.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.12.0
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
markdown-it-py==4.0.0
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.81.0
packaging==25.0
pandas==2.2.3
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.1
postgrest==2.27.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.1
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
realtime==2.27.0
requests==2.32.3
rich==14.2.0
scikit-learn==1.5.2
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==0.41.3
storage3==2.27.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.0
supabase-auth==2.27.0
supabase-functions==2.27.0
supafunc==0.3.3
tenacity==9.1.2
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
uvicorn==0.34.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
